feat: entities/feed/model/types.ts를 common.type, report.type, write.type ts파일로 분리
