+++ src/entities/feed/model/common.type.ts
@@ -19,16 +19,4 @@ export interface IFeed {
   updatedAt: string;
 }
 
-export interface IReportFeedReqDTO {
-  category: string;
-  content: string;
-  isBlind: boolean;
-}
-
-export interface IWriteFeedReqDTO {
-  content: string;
-  images: string[];
-  scope: TFeedScope;
-}
-
 export type TFeedScope = 'public' | 'friends' | 'private';
