+++ src/entities/feed/model/index.ts
@@ -1,1 +1,3 @@
-export * from './types';
+export * from './common.type';
+export * from './report.type';
+export * from './write.type';
