+++ src/entities/feed/model/report.type.ts
@@ -1,0 +1,5 @@
+export interface IReportFeedReqDTO {
+  category: string;
+  content: string;
+  isBlind: boolean;
+}
