+++ src/entities/feed/model/write.type.ts
@@ -1,0 +1,7 @@
+import { TFeedScope } from './common.type';
+
+export interface IWriteFeedReqDTO {
+  content: string;
+  images: string[];
+  scope: TFeedScope;
+}
